+++ pdmt5/__init__.py
@@ -2,12 +2,15 @@
 
 from importlib.metadata import version
 
-from .manipulator import Mt5Config, Mt5DataClient, Mt5RuntimeError
+from .exception import Mt5RuntimeError
+from .manipulator import Mt5Config, Mt5DataClient
+from .printer import Mt5DataPrinter
 
 __version__ = version(__package__) if __package__ else None
 
 __all__ = [
     "Mt5Config",
     "Mt5DataClient",
+    "Mt5DataPrinter",
     "Mt5RuntimeError",
 ]
