+++ pdmt5/exception.py
@@ -1,0 +1,7 @@
+"""MetaTrader5 specific exceptions."""
+
+from __future__ import annotations
+
+
+class Mt5RuntimeError(RuntimeError):
+    """MetaTrader5 specific error."""
